backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
httpx>=0.27.0
pydantic>=2.8.0
python-dotenv>=1.0.1

SQLAlchemy>=2.0.0
psycopg>=3.1.0
alembic>=1.13.0

python-jose[cryptography]>=3.3.0
email-validator>=2.2.0

# bcrypt：用于密码哈希（注意：不要装到 5.x 与旧版 passlib 混用会踩兼容性问题）
bcrypt==4.1.3
